# Extract expressions hash to reduce run time by 5-20%