Will you take an easy speedup?

```bash
bench "ruby --disable=gems src/rubyfmt.rb ~/.gems/activerecord-5.2.2.1/lib/active_record/reflection.rb"
```

**Before**
```
time                 258.7 ms   (255.6 ms .. 260.4 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 256.8 ms   (253.4 ms .. 258.4 ms)
std dev              3.020 ms   (580.4 μs .. 4.400 ms)
variance introduced by outliers: 16% (moderately inflated)
```

**After**
```
time                 210.6 ms   (203.2 ms .. 216.9 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 215.5 ms   (212.3 ms .. 219.3 ms)
std dev              4.790 ms   (3.543 ms .. 6.407 ms)
variance introduced by outliers: 14% (moderately inflated)
```

TL;DR: On a hairy Rails file, 258.7 ms => 210.6 ms